+++ src/js/lib/card-nicknames/stx.ts
@@ -1,10 +1,34 @@
 // Strixhaven (STX)
 export default [
   {
+    // https://scryfall.com/card/stx/36/would-you-like-to-buy-a-face%3F
+    setCode: "stx",
+    collectorNumber: "36",
+    realName: ["Arcane Subtraction"],
+    nickname: ["Would you like to buy a face?"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/38/croaktara
+    setCode: "stx",
+    collectorNumber: "38",
+    realName: ["Burrog Befuddler"],
+    nickname: ["Croaktara"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/42/library-strix
+    setCode: "stx",
+    collectorNumber: "42",
+    realName: ["Dream Strix"],
+    nickname: ["Library Strix"],
+    source: "Scryfall Preview Name",
+  },
+  {
     // https://scryfall.com/card/stx/63/a-poet-that-knows-it
     setCode: "stx",
     collectorNumber: "63",
-    realName: [],
+    realName: ["Arrogant Poet"],
     nickname: ["A Poet That Knows It"],
     source: "Scryfall Preview Name",
   },
@@ -12,31 +36,31 @@ export default [
     // https://scryfall.com/card/stx/64/sinister-domain
     setCode: "stx",
     collectorNumber: "64",
-    realName: [],
+    realName: ["Baleful Mastery"],
     nickname: ["Sinisiter Domain"],
-    source: "Scryfall Preview Name",
-  },
-  {
-    // https://scryfall.com/card/stx/68/deep-disappointment
-    setCode: "stx",
-    collectorNumber: "68",
-    realName: [],
-    nickname: ["Deep Disappointment"],
     source: "Scryfall Preview Name",
   },
   {
     // https://scryfall.com/card/stx/70/eyebat
     setCode: "stx",
     collectorNumber: "70",
-    realName: [],
+    realName: ["Eyetwitch"],
     nickname: ["Eyebat"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/81/into-the-unknown-chatroom
+    setCode: "stx",
+    collectorNumber: "81",
+    realName: ["Plumb the Forgotten"],
+    nickname: ["Into the Unknown Chatroom"],
     source: "Scryfall Preview Name",
   },
   {
     // https://scryfall.com/card/stx/91/academic-debate
     setCode: "stx",
     collectorNumber: "91",
-    realName: [],
+    realName: ["Academic Dispute"],
     nickname: ["Academic Debate"],
     source: "Scryfall Preview Name",
   },
@@ -44,7 +68,7 @@ export default [
     // https://scryfall.com/card/stx/92/forum-mooderator
     setCode: "stx",
     collectorNumber: "92",
-    realName: [],
+    realName: ["Ardent Dustspeaker"],
     nickname: ["Forum Mooderator"],
     source: "Scryfall Preview Name",
   },
@@ -52,23 +76,15 @@ export default [
     // https://scryfall.com/card/stx/93/geist-commander
     setCode: "stx",
     collectorNumber: "93",
-    realName: [],
+    realName: ["Blood Age General"],
     nickname: ["Geist Commander"],
-    source: "Scryfall Preview Name",
-  },
-  {
-    // https://scryfall.com/card/stx/94/conspiracy-theorist
-    setCode: "stx",
-    collectorNumber: "94",
-    realName: [],
-    nickname: ["Conspiracy Theorist"],
     source: "Scryfall Preview Name",
   },
   {
     // https://scryfall.com/card/stx/96/i-have-a-dragon-and-dragon-says-no-do-you-have-a-dragon%3F-didnt-think-so
     setCode: "stx",
     collectorNumber: "96",
-    realName: [],
+    realName: ["Draconic Intervention"],
     nickname: [
       "I Have a Dragon and Dragon Says 'No'. Do You Have a Dragon? Didn't Think So",
     ],
@@ -78,15 +94,31 @@ export default [
     // https://scryfall.com/card/stx/114/well-do-it-live!
     setCode: "stx",
     collectorNumber: "114",
-    realName: [],
+    realName: ["Start from Scratch"],
     nickname: ["We'll Do It Live!"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/121/bayou-groff
+    setCode: "stx",
+    collectorNumber: "121",
+    realName: ["Bayou Groff"],
+    nickname: ["Bayou Snapper"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/126/devouring-tendrils
+    setCode: "stx",
+    collectorNumber: "126",
+    realName: ["Devouring Tendrils"],
+    nickname: ["Hungry Tentacles"],
     source: "Scryfall Preview Name",
   },
   {
     // https://scryfall.com/card/stx/132/nightmares-beverage
     setCode: "stx",
     collectorNumber: "132",
-    realName: [],
+    realName: ["Fortifying Draught"],
     nickname: ["Nightmare's Beverage"],
     source: "Scryfall Preview Name",
   },
@@ -94,40 +126,56 @@ export default [
     // https://scryfall.com/card/stx/135/hungry-chomp-rider
     setCode: "stx",
     collectorNumber: "135",
-    realName: [],
+    realName: ["Karok Wrangler"],
     nickname: ["Hungry Chomp Rider"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/136/fern-guy-ly
+    setCode: "stx",
+    collectorNumber: "136",
+    realName: ["Leyline Invocation"],
+    nickname: ["Fern Guy-ly"],
     source: "Scryfall Preview Name",
   },
   {
     // https://scryfall.com/card/stx/138/master-of-minecraft
     setCode: "stx",
     collectorNumber: "138",
-    realName: [],
+    realName: ["Master Symmetrist"],
     nickname: ["Master of Minecraft"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/166/blood-researcher
+    setCode: "stx",
+    collectorNumber: "166",
+    realName: ["Blood Researcher"],
+    nickname: ["Blood Sommelier"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/177/the-scientific-bap
+    setCode: "stx",
+    collectorNumber: "177",
+    realName: ["Decisive Denial"],
+    nickname: ["The Scientific Bap"],
     source: "Scryfall Preview Name",
   },
   {
     // https://scryfall.com/card/stx/181/tiger-surfer
     setCode: "stx",
     collectorNumber: "181",
-    realName: [],
+    realName: ["Elemental Expressionist"],
     // previously named "Spirit Shaper" and "Storm Shaper" before settling on "Tiger Surfer"
     nickname: ["Tiger Surfer"],
-    source: "Scryfall Preview Name",
-  },
-  {
-    // https://scryfall.com/card/stx/183/elemental-calling
-    setCode: "stx",
-    collectorNumber: "183",
-    realName: [],
-    nickname: ["Elemental Calling"],
     source: "Scryfall Preview Name",
   },
   {
     // https://scryfall.com/card/stx/186/sublime-rehearsal
     setCode: "stx",
     collectorNumber: "186",
-    realName: [],
+    realName: ["Expressive Iteration"],
     nickname: ["Sublime Rehearsal"],
     source: "Scryfall Preview Name",
   },
@@ -135,24 +183,48 @@ export default [
     // https://scryfall.com/card/stx/187/frog-fractals
     setCode: "stx",
     collectorNumber: "187",
-    realName: [],
+    realName: ["Fractal Summoning"],
     nickname: ["Frog Fractals"],
     source: "Scryfall Preview Name",
   },
   {
-    // https://scryfall.com/card/stx/202/malestrom-muse
+    // https://scryfall.com/card/stx/208/needling-drake
     setCode: "stx",
-    collectorNumber: "202",
-    realName: [],
-    nickname: ["Malestrom Muse"],
+    collectorNumber: "208",
+    realName: ["Needlethorn Drake"],
+    nickname: ["Needling Drake"],
     source: "Scryfall Preview Name",
   },
   {
-    // https://scryfall.com/card/stx/209/oggyar-battle-seer
+    // https://scryfall.com/card/stx/234/silverquill-aristocant
     setCode: "stx",
-    collectorNumber: "209",
-    realName: [],
-    nickname: ["Oggyar Battle Seer"],
+    collectorNumber: "234",
+    realName: ["Silverquill Silencer"],
+    nickname: ["Silverquill AristoCAN'T"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/237/silverquill-bully
+    setCode: "stx",
+    collectorNumber: "237",
+    realName: ["Spiteful Squad"],
+    nickname: ["Silverquill Bully"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/257/reflective-hulk
+    setCode: "stx",
+    collectorNumber: "257",
+    realName: ["Reflective Golem"],
+    nickname: ["Reflective Hulk"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/stx/258/spell-satchel
+    setCode: "stx",
+    collectorNumber: "258",
+    realName: ["Spell Satchel"],
+    nickname: ["Magical Jansport"],
     source: "Scryfall Preview Name",
   },
 ];