+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Card Nicknames on Card Display Page
   - Fix issue where nickname would display before real name was known
+  - Add Strixhaven nicknames
 
 # 0.12.0